long for nanotime is fine as long as we dont need fractions

+++ bundles/binding/org.openhab.binding.alarmdecoder/src/main/java/org/openhab/binding/alarmdecoder/internal/model/ADZone.java
@@ -18,7 +18,7 @@ package org.openhab.binding.alarmdecoder.internal.model;
 public class ADZone {
 
     private int _id = 0;
-    private double _ts = 0;
+    private long _ts = 0;
     private boolean _state = false;
 
     /**
